Update dogstatsd client to v6.0.0 (#1143)

+++ test/Datadog.Trace.Tests/RuntimeMetrics/RuntimeEventListenerTests.cs
@@ -22,8 +22,8 @@ namespace Datadog.Trace.Tests.RuntimeMetrics
             listener.Refresh();
 
             statsd.Verify(s => s.Gauge(MetricsNames.ContentionTime, It.IsAny<double>(), 1, null), Times.Once);
-            statsd.Verify(s => s.Counter(MetricsNames.ContentionCount, It.IsAny<long>(), 1, null), Times.Once);
-            statsd.Verify(s => s.Gauge(MetricsNames.ThreadPoolWorkersCount, It.IsAny<int>(), 1, null), Times.Once);
+            statsd.Verify(s => s.Counter(MetricsNames.ContentionCount, It.IsAny<double>(), 1, null), Times.Once);
+            statsd.Verify(s => s.Gauge(MetricsNames.ThreadPoolWorkersCount, It.IsAny<double>(), 1, null), Times.Once);
         }
 
         [Fact]
@@ -53,12 +53,12 @@ namespace Datadog.Trace.Tests.RuntimeMetrics
                 mutex.Wait();
             }
 
-            statsd.Verify(s => s.Gauge(MetricsNames.Gen0HeapSize, It.IsAny<ulong>(), It.IsAny<double>(), null), Times.AtLeastOnce);
-            statsd.Verify(s => s.Gauge(MetricsNames.Gen1HeapSize, It.IsAny<ulong>(), It.IsAny<double>(), null), Times.AtLeastOnce);
-            statsd.Verify(s => s.Gauge(MetricsNames.Gen2HeapSize, It.IsAny<ulong>(), It.IsAny<double>(), null), Times.AtLeastOnce);
-            statsd.Verify(s => s.Gauge(MetricsNames.LohSize, It.IsAny<ulong>(), It.IsAny<double>(), null), Times.AtLeastOnce);
+            statsd.Verify(s => s.Gauge(MetricsNames.Gen0HeapSize, It.IsAny<double>(), It.IsAny<double>(), null), Times.AtLeastOnce);
+            statsd.Verify(s => s.Gauge(MetricsNames.Gen1HeapSize, It.IsAny<double>(), It.IsAny<double>(), null), Times.AtLeastOnce);
+            statsd.Verify(s => s.Gauge(MetricsNames.Gen2HeapSize, It.IsAny<double>(), It.IsAny<double>(), null), Times.AtLeastOnce);
+            statsd.Verify(s => s.Gauge(MetricsNames.LohSize, It.IsAny<double>(), It.IsAny<double>(), null), Times.AtLeastOnce);
             statsd.Verify(s => s.Timer(MetricsNames.GcPauseTime, It.IsAny<double>(), It.IsAny<double>(), null), Times.AtLeastOnce);
-            statsd.Verify(s => s.Gauge(MetricsNames.GcMemoryLoad, It.IsAny<uint>(), It.IsAny<double>(), null), Times.AtLeastOnce);
+            statsd.Verify(s => s.Gauge(MetricsNames.GcMemoryLoad, It.IsAny<double>(), It.IsAny<double>(), null), Times.AtLeastOnce);
             statsd.Verify(s => s.Increment(MetricsNames.Gen2CollectionsCount, 1, It.IsAny<double>(), compactingGcTags), Times.AtLeastOnce);
         }
